fix(difftest): move DiffState to separate files
This improves the code readability.

We also rename the generated DiffTestState filename to avoid naming
conflicts. We understand the current names are confusing, but we will
leave the naming issues to a future commit.

We also removed the undefined `display_commit_count` and
`display_commit_instr` functions.
