+++ src/test/csrc/difftest/diffstate.cpp
@@ -1,0 +1,45 @@
+/***************************************************************************************
+* Copyright (c) 2020-2025 Institute of Computing Technology, Chinese Academy of Sciences
+* Copyright (c) 2025 Beijing Institute of Open Source Chip
+*
+* DiffTest is licensed under Mulan PSL v2.
+* You can use this software according to the terms and conditions of the Mulan PSL v2.
+* You may obtain a copy of Mulan PSL v2 at:
+*          http://license.coscl.org.cn/MulanPSL2
+*
+* THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
+* EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
+* MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
+*
+* See the Mulan PSL v2 for more details.
+***************************************************************************************/
+
+#include "diffstate.h"
+#include "spikedasm.h"
+
+void DiffState::display() {
+  Info("\n============== Commit Group Trace (Core %d) ==============\n", coreid);
+  int group_index = 0;
+  while (!retire_group_queue.empty()) {
+    auto retire_group = retire_group_queue.front();
+    auto pc = retire_group.first;
+    auto cnt = retire_group.second;
+    retire_group_queue.pop();
+    Info("commit group [%02d]: pc %010lx cmtcnt %d%s\n", group_index, pc, cnt,
+         retire_group_queue.empty() ? " <--" : "");
+    group_index++;
+  }
+
+  Info("\n============== Commit Instr Trace ==============\n");
+  int commit_index = 0;
+  while (!commit_trace.empty()) {
+    CommitTrace *trace = commit_trace.front();
+    commit_trace.pop();
+    trace->display_line(commit_index, use_spike, commit_trace.empty());
+    commit_index++;
+  }
+
+  fflush(stdout);
+}
+
+DiffState::DiffState(int coreid) : use_spike(spike_valid()), coreid(coreid) {}
